+++ lib/components/fo_modal/fo_modal_component.dart
@@ -7,10 +7,9 @@ import 'package:fo_components/components/fo_button/fo_button_component.dart';
     selector: 'fo-modal',
     styleUrls: ['fo_modal_component.css'],
     templateUrl: 'fo_modal_component.html',
-    directives: [coreDirectives, FoButtonComponent])
+    directives: [coreDirectives, FoButtonComponent],
+    changeDetection: ChangeDetectionStrategy.OnPush)
 class FoModalComponent implements OnDestroy {
-  FoModalComponent();
-
   @override
   void ngOnDestroy() {
     _onVisibleChangeController.close();