+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 2.1.0
+* FoModal, FoTextInput, FoTextareaInput, FoDropdownSelect now has OnPush changeDetection
+* Fixed issue where dropdown didn't update height properly when options contains images (see https://github.com/afpatmin/fo_components/issues/78)
+
 # 2.0.1+1
 * Added attribute FoTextInput.autocomplete
 * Updated some dependencies in preparation for null safety